+++ interface/build.gradle
@@ -12,5 +12,6 @@ repositories {
 dependencies {
     compile 'com.google.guava:guava:23.6-jre'
     compile 'com.google.code.gson:gson:2.8.2'
+    compile 'io.netty:netty-all:4.1.19.Final'
     testCompile group: 'junit', name: 'junit', version: '4.12'
 }
